// Frontend/src/views/PP/enums/pp009.ts
export enum PP009Status {
  Draft = 'Draft',
  Edit = 'Edit',
  WaitingApproval = 'WaitingApproval',
  Approved = 'Approved',
  Rejected = 'Rejected',
}

export enum ContractInvitationHeader {
  Detail = 'Detail',
  InvitationDocument = 'InvitationDocument'
}

export enum PP009Accordion {
  Acceptor = 'Acceptor',
}